Update 02-patterns.mdx
Hello there! Just asking out of curiosity. How can the Page component have access to the params object and extract the username from it, if there is no dynamic route segment? I mean if the folder is not named [username] and then have a page.tsx file inside of THAT folder, how can it extract the username from the params object? I'm now learning Next.js 13, so I might be wrong about it. So please explain to me if I'm wrong. :)