fix: Don't reset shallow URL updates on prefetch
Between 14.0.2-canary.6 and 14.0.2-canary.7, a change was introduced
in vercel/next.js#56497 that turned the Redux store state into Promises,
rather than a synchronous state update.

This caused the `sync` function used to send state update to the Redux
Devtools to be recreated on every dispatch, which in turn, by referential
instability, caused the `HistoryUpdater` component to re-render and
trigger a `history.replaceState` with no particular change, but with the
internal `canonicalUrl`.

When an app does a soft/shallow navigation by calling history methods
directly (currently the only way to do shallow search params updates in
the app router), these changes would have been overwritten by any
prefetch (eg: hovering or mounting a Link), which is usually a no-op
for the navigation state.

This commit removes the `sync` function for the Redux devtools,
as it cannot function as-is: actual state updates need to be awaited
and forwarded somewhere else in the router if this behaviour is to
be maintained, and should be decoupled from history calls to prevent
side-effects.