Revert "Add cacheNodeSeedData to RSC payload (#58566)"
This reverts commit d5aea2c5cf462b3d1c31fb250c1f8b5eae3664fb, which
added an additional field to the RSC payload intended to represent a
tree of subtree data for all the nested layouts in a response. I didn't
think about the fact that the subtreeData slot, which represents the
node at the root of the subtree, would be included twice.

Instead, we should replace the subtreeData slot to be of the
CacheNodeSeedData tree type, instead of just a single node. I'll do that
in the following commits.