Lift up shouldSkipComponentTree check
I need to use the result of createComponentTree in two different parts
of the Flight response, but it shouldn't be called during a prefetch. So
I hoisted the shouldSkipComponentTree check to earlier in the control
flow. This also lets us get rid of an extra React element that was being
used to lazily call createComponentTree.