Pass CacheNodeSeedData instead of ReactNode
The Flight response type includes the component node for the root of the
subtree that was rendered. Instead of passing just the root node, we
want to pass the entire tree of nested layouts. As an imcremental step,
this changes the type to pass a CacheNodeSeedData object, but it passes
only a partial tree that includes only the root node. In the next step,
we'll pass the entire tree in this object, and in exchange we'll remove
the initialChildNode prop to LayoutRouter.