Delete preloadComponent module
We previously used preloadComponent to render nested page/layouts in
parallel, but now that all layouts are passed to React at the top level
(#58669), we no longer need this special module.

A flaw of preloadComponent was that it only called the top-most
component in the layout. In the new approach, React can render the
entire layout tree in parallel.