Remove `cache` field from action types
The app router state used to be managed by `useReducer`, so it was
written to be resilient to rebasing — the same action may be processed
multiple times. Now that we've lifted the router state outside of React
(#56497), each action runs only a single time. So we can simplify some
of the logic.

Previously, actions that update that cache were passed an empty cache
object as part of the action payload. Now, we can instead create these
objects on demand in the reducer.

(We should do the same for the `mutable` object, but there's a unit test
which relies on this implementation detail, so I've left that for a
separate PR.)